Rename lib/tracker_lqsr.py -> lib/tracker_lsqr.py
