+++ lib/tracker_lsqr.py
